Enhance CI coverage reporting and add integration tests for telemetry
- Updated the CI workflow to enforce a minimum line coverage of 80% for tests, providing detailed coverage summaries and error messages for unmet thresholds.
- Added a new `tests/telemetry.rs` file containing integration tests for the telemetry module, ensuring proper logging and field validation without sensitive information exposure.
- Revised the README.md to include coverage job details and its requirements for clarity.
