+++ README.md
@@ -312,6 +312,8 @@ cargo doc --all-features --no-deps --open
 
 **Security audits:** On **push/PR**, CI runs `rustsec/audit-check` and can post a **check** on the change. A **weekly** workflow (Mondays, UTC) runs plain **`cargo audit`** only — it does **not** open GitHub issues (least-privilege token); a **failed run** means the lockfile matches advisories you should address.
 
+**Coverage:** The **coverage** job enforces a **minimum line coverage** (currently **80%**). If it fails, the log and job summary print the measured percentage vs. the required minimum and an annotation explains that tests under `src/` need to cover more lines (run `cargo llvm-cov --all-features` locally to inspect).
+
 ### Run a single example
 
 ```bash
